pyproject.toml
[project]
name = "webp-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
    "pillow-heif>=1.1.1",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "isort>=7.0.0",
]
